requirements.txt
fastapi
uvicorn
roboflow
python-multipart
python-dotenv